Move deepDepSkip to interfaces
