+++ grafast/grafast/src/steps/listTransform.ts
@@ -13,8 +13,9 @@ import { withGlobalLayerPlan } from "../engine/lib/withGlobalLayerPlan.js";
 import type { GrafastError } from "../error.js";
 import type { ConnectionCapableStep, ExecutionDetails } from "../index.js";
 import type { GrafastResultsList } from "../interfaces.js";
+import { $$deepDepSkip } from "../interfaces.js";
 import type { ListCapableStep } from "../step.js";
-import { $$deepDepSkip, ExecutableStep, isListCapableStep } from "../step.js";
+import { ExecutableStep, isListCapableStep } from "../step.js";
 import { __ItemStep } from "./__item.js";
 
 export type ListTransformReduce<TMemo, TItemPlanData> = (
